// packages/ide/src/types/ai-hardware-profiling-types.ts
/**
 *
 * Copyright (c) 2026 Analog Devices, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */

export type BuildStatus = "idle" | "building" | "built" | "error";
export type DeployStatus =
  | "undeployed"
  | "deploying"
  | "running"
  | "stopped"
  | "error";

export type ApplicationStatus = {
  buildStatus: BuildStatus;
  deployStatus: DeployStatus;
};

export type HardwareResources = {
  usbPorts: string[];
  debuggers: string[];
  port?: string;
};

export type ProfilingConfiguration = {
  selectedUsbPort?: string;
  selectedDebugger?: string;
};